feat: fix controller status codes and add item field validations - Fix incorrect HTTP status codes in ItemController - Add Email Validation to Item.java - Add validation for name, description, and status fields in Item.java

+++ src/main/java/com/siemens/internship/ItemController.java
@@ -25,33 +25,40 @@ public class ItemController {
     @PostMapping
     public ResponseEntity<Item> createItem(@Valid @RequestBody Item item, BindingResult result) {
         if (result.hasErrors()) {
-            return new ResponseEntity<>(null, HttpStatus.CREATED);
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST); // return a bad request if validation fails
         }
-        return new ResponseEntity<>(itemService.save(item), HttpStatus.BAD_REQUEST);
+        return new ResponseEntity<>(itemService.save(item), HttpStatus.CREATED); // return a created response
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Item> getItemById(@PathVariable Long id) {
         return itemService.findById(id)
                 .map(item -> new ResponseEntity<>(item, HttpStatus.OK))
-                .orElse(new ResponseEntity<>(HttpStatus.NO_CONTENT));
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND)); // return NOT_FOUND if the item does not exist
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Item> updateItem(@PathVariable Long id, @RequestBody Item item) {
         Optional<Item> existingItem = itemService.findById(id);
         if (existingItem.isPresent()) {
-            item.setId(id);
-            return new ResponseEntity<>(itemService.save(item), HttpStatus.CREATED);
+            item.setId(id); // set the ID of the item to be updated
+            Item newItem = itemService.save(item); // save the updated item
+            return new ResponseEntity<>(newItem, HttpStatus.OK); // return the updated item with OK status
         } else {
-            return new ResponseEntity<>(HttpStatus.ACCEPTED);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND); // return NOT_FOUND if the item does not exist
+
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteItem(@PathVariable Long id) {
-        itemService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.CONFLICT);
+        Optional<Item> item = itemService.findById(id);
+        if (item.isPresent()) { // check if the item exists
+            itemService.deleteById(id); // delete the item
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // return NO_CONTENT status
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND); // return NOT_FOUND if the item does not exist
+        }
     }
 
     @GetMapping("/process")
